ch 4 blue gradient

+++ ch1/src/point.rs
@@ -1,3 +1,3 @@
 use nalgebra::Vector3;
 
-pub type Point = Vector3<i32>;
+pub type Point = Vector3<f32>;
